+++ api/operator/v1beta1/vmcluster_types.go
@@ -24,7 +24,7 @@ type VMClusterSpec struct {
 	// The default value is 1 (one month).
 	// See [retention](https://docs.victoriametrics.com/victoriametrics/single-server-victoriametrics/#retention) docs for details.
 	// +optional
-	// +kubebuilder:validation:Pattern:="^[0-9]+(h|d|y)?$"
+	// +kubebuilder:validation:Pattern:="^[0-9]+(h|d|w|y)?$"
 	RetentionPeriod string `json:"retentionPeriod,omitempty"`
 	// ReplicationFactor defines how many copies of data make among
 	// distinct storage nodes